feat: Export Like to Components
